refactor(web): extract route feature sections
Move Area, Project, and dashboard page sections into feature modules so routes stay focused on params, loading states, navigation, and composition.
