+++ apps/web/src/features/projects/components/project-lifecycle-actions.tsx
@@ -1,0 +1,119 @@
+import type { Doc } from "@convex/_generated/dataModel";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@vita-os/ui/components/alert-dialog";
+import { Button } from "@vita-os/ui/components/button";
+import { CheckCircle2, Trash2, XCircle } from "lucide-react";
+
+interface ProjectLifecycleActionsProps {
+  project: Doc<"projects">;
+  onStateChange: (state: "completed" | "dropped") => void;
+  onDelete: () => void;
+}
+
+export function ProjectLifecycleActions({
+  project,
+  onStateChange,
+  onDelete,
+}: ProjectLifecycleActionsProps) {
+  return (
+    <div className="flex items-center gap-2 border-t border-border/50 pt-6">
+      <AlertDialog>
+        <AlertDialogTrigger
+          render={<Button variant="outline" size="sm" className="gap-1.5" />}
+        >
+          <CheckCircle2 className="h-3.5 w-3.5" />
+          Complete
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Complete project?</AlertDialogTitle>
+            <AlertDialogDescription>
+              &ldquo;{project.name}&rdquo; will be marked as completed.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction onClick={() => onStateChange("completed")}>
+              Complete
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+
+      <AlertDialog>
+        <AlertDialogTrigger
+          render={
+            <Button
+              variant="ghost"
+              size="sm"
+              className="gap-1.5 text-muted-foreground"
+            />
+          }
+        >
+          <XCircle className="h-3.5 w-3.5" />
+          Drop
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Drop project?</AlertDialogTitle>
+            <AlertDialogDescription>
+              &ldquo;{project.name}&rdquo; will be marked as dropped.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={() => onStateChange("dropped")}
+              className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+            >
+              Drop
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+
+      <div className="flex-1" />
+
+      <AlertDialog>
+        <AlertDialogTrigger
+          render={
+            <Button
+              variant="ghost"
+              size="sm"
+              className="gap-1.5 text-muted-foreground hover:text-destructive"
+            />
+          }
+        >
+          <Trash2 className="h-3.5 w-3.5" />
+          Delete
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Delete project?</AlertDialogTitle>
+            <AlertDialogDescription>
+              &ldquo;{project.name}&rdquo; will be permanently removed.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={onDelete}
+              className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+            >
+              Delete
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </div>
+  );
+}
